+++ src/modern_yolonas/__init__.py
@@ -4,6 +4,7 @@ from modern_yolonas._version import __version__
 from modern_yolonas.model import YoloNAS
 from modern_yolonas.weights import load_pretrained, load_checkpoint, transfer_to
 from modern_yolonas.inference.detect import YoloNASDetector
+from modern_yolonas.inference.embed import FEATURE_LAYERS, YoloNASEmbedder
 from modern_yolonas.coco import COCOClass
 from modern_yolonas.inference.visualize import COCO_NAMES
 
@@ -42,6 +43,8 @@ __all__ = [
     "load_checkpoint",
     "transfer_to",
     "YoloNASDetector",
+    "YoloNASEmbedder",
+    "FEATURE_LAYERS",
     "COCOClass",
     "COCO_NAMES",
 ]